requirements.txt
pdfplumber==0.8.0
pymongo==4.3.3
argparse==1.4.0
Unidecode==1.3.6
beautifulsoup4==4.12.2
progressbar2==4.2.0
requests==2.31.0
black
psycopg2-binary